Updates/fixels/comments
The demosaic'd LMS array is now returned as well as the sRGB image.  Sometimes we want that.

I think there was a row/col bug in the setup and call to griddata, and also a scaling bug on the LMS data before the conversion to sRGB.  I am not 100% sure, so these are commented and I will post an issue in a second

Signed-off-by: David Brainard <brainard@psych.upenn.edu>