Modified sensorSet('cone type') so that it makes both the cone type and cfa.pattern fields the same.
This is required to keep things consistent.  It's done on sensorCreate('human') (that is, the two fields are made to match).

Probably it should also be done on sensorSet('pattern') and sensorSet('cfa').

Actually, probably there should not be a cone type field at all, or if there should I am not seeing why.

Signed-off-by: David Brainard <brainard@psych.upenn.edu>