sensorGet adaptation code
Changed 'adapted volts' -> 'adapted data' in help text, because that is what the routine expects.

In the adatped data calculation, there was a call to get 'adapted volts'.  This threw an error.  Since the line computed a variable named offset, we changed this to a call to get 'adaptation offset'.  Now it runs.  But we are not sure this is what was intended.