Revert "Fixed movie playing bug in conemosaic after changing the return order of plot() to conform with other plotting rtines"
This reverts commit 830a63f99db82be1026a0e9c75e796e42d018f6a.

I edited emSequence to change the integrationTime in the coneMosaic so that it iequals the number of frames times the step duration.
Fixed other little stuff that my commits seem to have broken.
The window mostly seems OK now.
I am going to move the t_rgcEyeMovement script out of the git push path for a while, because it changes all the time and it is annoying to see it keep getting updated.