Small speeding up by setting lineStyle to ‘none’
I agree that there is something going on with line plotting in the comeMosaicWindow.

(1) It is much slower than in a figure with no GUI controls.
(2) The lines are not smooth like in the figure with no GUI controls.

Another way to significantly increase the rendering speed is to set the dAngle to 60 (from 15) in line 60, which results in plotting hexagons instead of circles.

Signed-off-by: Nicolas Cottaris <cottaris@sas.upenn.edu>