I removed the coneCurrentSignal from the outerSegment objects
The current is returned and placed in the coneMosaic object current slot

I adjusted the time representation.  All time axes are now controlled as follows
   * The cone mosaic integrationTime sets the temporal sampling interval
   * The number of eye movement samples determines how many frames
These are the only time values for the cone mosaic.  When you compute an oiSequence, it can have its own time base and NC manages that in computeForOISequence.  If you calculate the osBioPhys model you can manage that in how you set the timeStep parameter.

I deprecated generateBioPhys....
I changed the returns on a number of functions to conform with the above.
I made various changes to the coneMosaicWindow for improved visibility and alignment of the images (e.g., by adding the color bar and handling gamma correctly).