A number of serious bugs removed.  Features added.
First, there was an error in how the photocurrent was calculated.  I was normalizing the impulse response function, and that was a mistake.  So, that is removed.

Second, there were issues in how the conv2() operation in computeCurrent was handled.  We need to be careful about which is the filter, which the data, and the relative lengths of the two.  I write a new scropt (s_matlabConv2.m) to explain what appears to be the right way to do things.  That is referenced in computeCurrent now and the script is self-explanatory.

I added various tests, such as an impulse response function test and the ability to return the temporall downsampled impulse response functions.  These can be compared with the IR functions on the 1 ms time base, and those plots come up in v_cmCurrentImpulse.m

Some small improvements in plotting and visualization.

Onward.