The comments were wrong in retinalLocationToTEE.m  I fixed them.  rho/theta were reversed and no units were specified.
I adjusted the comments and renamed key parameters in ir.m, clarifying which come from the bipolar input and which are set by the constructor.  The naming convention was changed to reflect the true function of some of the parameters (size, rather than spacing for example).

Many units and comments added.