Updated the noise flag from 0 to 'none' in many scripts. Also fixed various other noise flag related problems.
Changed the tCenter and tSurround to be a single temporal impulse response function rather than a different one with different weights for each of the 'RGB' inputs to the ganglion cell mosaic object.  That is still a work in progress.

Ran ieValidationFullAll a few more times and fixed some of the scripts that had either noise flag or other problems.