Got the display color correction validation to work
There was a slight difference in the isomerizations once I dealt with the change in cone collection area, but this was on the order of 20 for isomerizations on the order of 1000-7000 and I attribute it to some different noise draw in the new code, for reasons that aren’t completely apparent to me.  But I decided to re-write the validation file and charge onwards.

Signed-off-by: David Brainard <brainard@psych.upenn.edu>