Documentation for today
Starting on a long march through the tutorials

Adding Contents.m as I go.  If a tutorial has a line in a Contents.m file, then I have reviewed it and/or tried to make it better.

Things hanging in tutorials that are noted as reviewed are flagged by [DHB NOTE: …] in comments.  Global search should find all of these.

Renamed tutorial subdirs with t_, so that help and doc are a little smarter about what they print out.  Need to do this with validation subdirs as well.

Signed-off-by: David Brainard <brainard@psych.upenn.edu>