Allow ‘eccentricityDegs’ key/value pair to osBioPhys
This seemed like a no brainer.  I wanted it as soon as I tried reading one of the tutorials, and when I went to look at osBioPhys there were comments there that this should happen.

I did it so that it is backwards compatible (for now), but so that a cleaner call with ‘eccentricityDegs’ is respected.

Signed-off-by: David Brainard <brainard@psych.upenn.edu>