Change compute -> underDevelopment_wavefront
The reason the examples don’t work is because these routines are not finished.  Changed directory name to make this clear, and also added a note in each function and throw an error at the top of each function, to make this status clear.

These are not currently used anywhere that is not also marked as under development.