Take key value/pairs
Previously, this took no inputs and simply set things up with default values.  The documentation indicated, however, that it took key value pairs.  I rewrote so that it does take key/value pairs.

As part of this, I implemented into ieParamFormat the ability to take a cell array, in which case all the odd entries (the keys) are converted to ieParam format.

Not yet fully tested.

Need to go back and deal with all the examples, which previously ignored any pass key/value pairs.