Change wvfCreate(‘wave’, …) -> wvfCreate(‘calc wavelengths’,…)
The latter is what the documentation says will work, and indeed I recoded wvfCreate using the input parser so that ‘wave’ no longer works.  Note that there are two different wavelengths that the wvf structure maintains, one is that used to measure the wavefront and the other is those at which the psf is calculated.  So ‘wave’ is not so good as a synonym here.