Numerous changes, but most important was to default for um per degree
The default for um per degree was set to 330 in this routine, which is quite different from the 300 we used elsewhere.  I added a field and made it set and getable, and made the default 300.

This broke a lot of validations, and I have now updated the corresponding validation files.  It will eventually improve other things, I hope.