Allow setting of umPerDegree via varargin for opticsCreate, for ‘wvfhuman’
Also accept defaults if any of the extra args are passed as empty.

This still has a bad bug, namely that 3mm Thibos data is loaded no matter what is passed as the pupil diameter in mm.  That will lead to problems if any pupil diameter other than 3mm is passed.  But I’m not fixing that now, because I am trying to fix some validations in IBIOColorDetect and can’t change too much at once.