Use ifftshift rather than circshift to put sf (0,0) in the upper right of otf
The wvf object keeps sf (0,0) in the center of the otf it returns.  But the oi/optics convention is for sf (0,0) to be in the upper right.  I think ifftshift is the way to make this shift, and am trying to do that everywhere.

It’s possible that we should change one of these conventions so that they are consistent across wvf and oi/optics, but one step at a time.