Use PTB PsfToOtf to convert psf to otf
My goal is to localize all the psf <-> conversion to PTB routines PsfToOtf and OtfToPsf.  That will keep everything consisten, up to the convention of where we keep the center/dc position in the matrix.  And that we know how to move around in a manner consistent with the behavior of the PTB routines.