Added an ifftshift before fft.
This doesn’t affect the output, because we take the intensity and discard the phase.  But I think it is more consistent with Matlab’s fft conventions.

Also added some check code that can be turned on.  The various checks all made sense when I checked them.

These changes were part of an attempt to sort out even/odd support effects, but turned out not to matter for that.