Turn off warning 'MATLAB:Figure:RecursionOnClose' in ieInit
This prevents a printout of warnings of this general form when we close isetbio windows.

Warning: A callback recursively calls CLOSE. Use DELETE to prevent this message.

In close>request_close_helper
In close>request_close
In close
In close>request_close
In close
In t_conesMosaicBasic (line 12)
In UnitTest.runProjectTutorials>runTheTutorial (line 97)
In UnitTest.runProjectTutorials (line 49)
In ieRunTutorialsAll (line 30)