Implement suggestion in issue 127
The issue text is below.  This has been hanging for 1.5 years, I went ahead an implemented the idea (divide spds by 5) and saved out the corrected file as CRT-Sony-HorwitzLab.  Deleted the version with “NOT RIGHT” at the end.

****
I computed the luminance of the white point of the described display and it exceeds 800 cd/m2. This is not plausible for a CRT.

I think what probably happened is that Greg provided phosphor spectra as they are typically measured in the PTB, which is in units of Watts/[m2-sr-wavelengthBand], and that these were then imported directly into isetbio, whose standard units are Watts/[m2-sr-nm]. Because the wavelength bands were 5 nm wide, this leads to the numbers being a factor of 5 too big.

The fix is to divide the phosphor spds by 5, if this is what happened.

I was too chicken, and simply renamed the file with the postpended -NOTRIGHT.