All examples (“% Examples:” followed by block comment run or intentinally skipped.
First pass on examples complete.  They all either run or are explicitly skipped.

Some of the skips are good — examples that require user input and that we don’t want to autorun, or that have side effects that we don’t want to autorun.

Some of the skips are bad — broken examples that I could not figure out how to fix easily.  Will work on an issue that lists these.