Fix to make dlMTF examples work when run from the full example script.
It turns out that a call to oiGet broke if there was a window open with a scene in it, because a call to vcGetSelectedObject was called to get a scene when none was passed, and that call returned a number rather than a scene because the first rather than the second output arg was picked up.

I changed the call to vcGetObject, which seems to have fixed the problem.

The behavior was hard to track down because the examples ran fine when no scene window was open, but one is opened up by some previous example when they are all run from the scriptl.

Also deleted comment that example in dlMTF is broken.