Cosmetic changes to comments.
a) It’s nice not to have blank lines in header comments - the blank will truncate the printout when you use Matlab’s help funciton.

b) Moved example to a block comment, below a line that says % Examples.  This allows the examples to be selected and executed, and also lets them be autorun by ieRunExamplesAll (see ExampleTestToolbox).