Two new files.  psf2zcoeff and s_opticsPSF2Zcoeffs.  These are demonstrations of estimating the Zernike polynomial coefficients from a point spread function.
The other files were only touched for comments and to add zpupildiameter and zwavelength as options into wvfGet.  The reason is for compatibility with ISETCam.  I wanted to refer to the 'measured' parameters using z to indicate that these parameters are attached to the conditions under which the Zernike polynomial coefficients are measured.  I thought the z would help me more than saying 'measured'.