Fix interface between PTB and ISETBio power conventions, going between PTB cal and ISETBio display
The PTB and ISETBio worlds use a different convention for specifying power in spectra.  PTB uses power per wl band, ISETBio power per nm.  Some care is needed converting spectra when going from one world to the other, and this was not being handled in the routines here. Indeed, it was not even being handled consitently on two threads through GenerateIsetbioDisplayFromPTBCalStruct, because a call to SplineSpd adjusts spectrum values to maintain the power per wlband convention, but this adjustment was not accounted for.

Fixed this up so it's all consistent, I think.  And added a boatload of comments to these routines.