For human eye ISET3d rendering - accommodation in filename
I added the lenses for now so we can use them with ISET3d.  If I make the architectural change to base on ISETCam, then we can eliminate them.

Added the accommodation in the filename.

Modified the piWRS() method to call the main piWRS function, though we wrap it because of the use pinhole flag.

Added ieNotDefined.  See architecture comment above.