The importance of creating the proper OI
We have to set an OI that is 'human wvf' to work correctly with the cone mosaics (both cMosaic and coneMosaicRect).  If the f-number is not right, the spatial scale of the OI will not match the expected spacing of the cone receptors.

We need to understand and explain this better.  But this means that the ISETBio calculations should be using oiCreate('human wvf').  Just oiCreate; will produce an OI with the wrong fnumber.