Merged ISETBio/isettools/human into ISETCam/human
These are mainly older human calculations that I still use as part of the ISETCam teaching (psych 221).  By leaving them in ISETCam, the students won't have to download ISETBio.

The ISETBio calculations are all much more sophisticated than this.  And they will always be available to ISETBio in any event.