Turn off fancy parpool control by default.
Turned off fancy parpool control by default, with a variable that allows the user to turn it back on.

It takes a long time to start and restore parpools.  We want tutorials to be quick and nimble by default.  Better strategy is to set default parameters so that the tutorial doesn't crush a typical machine.

I put in a comment that the user can try turning this on if there are strange parpool crashes, but I think we have now reduced the scale of these tutorials so it is OK.

If we discover that these tutorials are regularly crashing, we can change the default.