Making progress on scaling the bitmap E, which was fixed in pixel size previously.
The way it was scaled in our chromatic aberration experiments was to vary the DPI of a fixed size image.  This caused the E to scale on the retina, but also the size of the background scaled with the size of the E.  That should not have affected the calculations, but it confused me so I am trying to rescale the bitmap instead.  Getting there.