Get cone qe properly interpolated out past the wavelength range on which the table is defined
We need to linearly interpolate the log absorbance out past the defined wavelength range, so special case when you are out past the long end of the table and handle this.  We need the cone fundamentals out to long wavelengths for AO modeling.

Also deleted a lot of stray vertical blank space as I came across it.