Fix this to map 0 to 0, not some small number.
This problem arsises because the ie gamma corrections behave strangely, or perhaps we don’t understand how to call them.

Rewrite this routine to ensure that when you ask for zero linear rgb that is really what you get.  And also to not allow return values greater than 1 for RGB.

Probably should carefully check the ie gamma correction routines, or switch to the PTB equivalents, but life is short.